OLD:
```
Benchmark_Ctx_IPs-8      5335536               221 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      5490664               219 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      5296419               218 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      5501774               227 ns/op              48 B/op          1 allocs/op
```

NEW:
```
Benchmark_Ctx_IPs-8      6307110               187 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      6372991               190 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      6499392               188 ns/op              48 B/op          1 allocs/op
Benchmark_Ctx_IPs-8      6374119               188 ns/op              48 B/op          1 allocs/op
```